Adding zlibx (workaround for zlib termination problem)
For some reason DEFLATE streams produced by 7-Zip are somehow terminated
in a way that D's zlib does not handle well. inflate keeps returning S_OK,
but won't write anything to the output buffer. zlibx contains a function,
which always decompresses the exact number of bytes (which is always known
in our cases). The downside is that if the compressed stream has more data
than what we expected, it will be silently truncated.